requirements.txt
# Whisper AI Transcription Dependencies
openai-whisper>=20240930
torch>=2.0.0
torchaudio>=2.0.0
ffmpeg-python>=0.2.0
numpy>=1.21.0
tqdm>=4.60.0
regex>=2022.1.18
tiktoken>=0.3.0

# Gemini Image Generation Dependencies
google-genai>=1.37.0
python-dotenv>=1.1.1
